Update cipher spelling
