+++ test/enigma_test.rb
@@ -1,1 +1,2 @@
-require "minitest/autorun"
+require "./test/test_helper"
+require "./lib/enigma"
