+++ test/test_helper.rb
@@ -1,0 +1,3 @@
+require "simplecov"
+SimpleCov.start
+require "minitest/autorun"
